+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   Font class included in public API.
+-   Glyphmap with dynamic font demo.
+
 ## [0.2.0] - 2022-01-08
 
 ### Added