+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.3.0] - 2022-01-9
+
 ### Added
 
 -   Font class included in public API.
@@ -39,6 +41,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   Initial release with on the fly glyph texture generation, glyphmap gameobject, regression tests, & a deployed demo.
 
-[unreleased]: https://github.com/agogpixel/phaser3-glyph-plugin/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/agogpixel/phaser3-glyph-plugin/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/agogpixel/phaser3-glyph-plugin/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/agogpixel/phaser3-glyph-plugin/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/agogpixel/phaser3-glyph-plugin/releases/tag/v0.1.0